+++ api/tests/test_importer.py
@@ -8,8 +8,8 @@ from app.all import import_all_from_env
 
 
 @mock.patch("app.plugins.collection", return_value=Mock())
-@mock.patch("app.all.import_default_config_from_env", return_value=None)
-@mock.patch("app.all.import_datasources_from_env", return_value=None)
+@mock.patch("app.config_repo.import_default_config_from_env", return_value=None)
+@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_plugins_from_env(mock_datasources, mock_plugins, mock_collection):
     os.environ[
@@ -30,8 +30,8 @@ async def test_import_plugins_from_env(mock_datasources, mock_plugins, mock_coll
 
 @mock.patch("app.config_repo.collection", return_value=Mock())
 @mock.patch("app.config_repo.get", return_value=None)
-@mock.patch("app.all.import_plugins_from_env", return_value=None)
-@mock.patch("app.all.import_datasources_from_env", return_value=None)
+@mock.patch("app.plugins.import_plugins_from_env", return_value=None)
+@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_feature_flags_from_env(
     mock_datasources, mock_plugins, mock_get, mock_collection
@@ -66,8 +66,8 @@ async def test_import_all_from_env_with_broken_plugin(mock_collection):
 
 @mock.patch("app.config_repo.collection", return_value=Mock())
 @mock.patch("app.config_repo.get")
-@mock.patch("app.all.import_plugins_from_env", return_value=None)
-@mock.patch("app.all.import_datasources_from_env", return_value=None)
+@mock.patch("app.plugins.import_plugins_from_env", return_value=None)
+@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_feature_flags_from_env_update(
     mock_datasource, mock_plugins, mock_get, mock_collection
@@ -92,8 +92,8 @@ async def test_import_feature_flags_from_env_update(
 
 @mock.patch("app.config_repo.collection", return_value=Mock())
 @mock.patch("app.config_repo.get")
-@mock.patch("app.all.import_plugins_from_env", return_value=None)
-@mock.patch("app.all.import_datasources_from_env", return_value=None)
+@mock.patch("app.plugins.import_plugins_from_env", return_value=None)
+@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_feature_flags_from_env_no_insert_or_update(
     mock_datasource, mock_plugins, mock_get, mock_collection
@@ -114,8 +114,8 @@ async def test_import_feature_flags_from_env_no_insert_or_update(
 
 
 @mock.patch("app.datasource.datasource_repo", return_value=Mock())
-@mock.patch("app.all.import_plugins_from_env", return_value=None)
-@mock.patch("app.all.import_default_config_from_env", return_value=None)
+@mock.patch("app.plugins.import_plugins_from_env", return_value=None)
+@mock.patch("app.config_repo.import_default_config_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_datasources_from_env(mock_plugins, mock_config, mock_datasource):
     os.environ[
